+++ sdk/internal/test/tracing/validator_test.go
@@ -5,6 +5,7 @@ package tracing
 
 import (
 	"context"
+	"errors"
 	"testing"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
@@ -14,55 +15,69 @@ import (
 
 func ExampleNewSpanValidator() {
 	t := &testing.T{}
+
+	// attributes and links used when starting the span
+	initialAttr := tracing.Attribute{Key: "initialAttrKey", Value: "initialAttrValue"}
+	initialLink := tracing.Link{Attributes: []tracing.Attribute{{Key: "initialLinkKey", Value: "initialLinkValue"}}}
+	// attributes and links added after starting the span
+	testAttr := tracing.Attribute{Key: "testSetAttrKey", Value: "testSetAttrValue"}
+	testLink := tracing.Link{Attributes: []tracing.Attribute{{Key: "testAddLinkKey", Value: "testAddLinkValue"}}}
+
 	// create a span validator that will verify spans with the given name, kind, status, attributes, and links
+	// the span matcher will verify the span at the end of the tests
 	provider := NewSpanValidator(t, SpanMatcher{
-		Name:   "TestSpan",
-		Kind:   tracing.SpanKindClient,
-		Status: tracing.SpanStatusUnset,
-		Attributes: []tracing.Attribute{
-			{Key: "testKey", Value: "testValue"},
-		},
+		Name:       "TestSpan",
+		Kind:       tracing.SpanKindClient,
+		Status:     tracing.SpanStatusUnset,
+		Attributes: []tracing.Attribute{initialAttr, testAttr},
+		Links:      []tracing.Link{initialLink, testLink},
 	})
 	tracer := provider.NewTracer("module", "version")
 
+	// start a test span with initial attributes and links
 	ctx, endSpan := runtime.StartSpan(context.Background(), "TestSpan", tracer, &runtime.StartSpanOptions{
-		Kind: tracing.SpanKindClient,
-		Attributes: []tracing.Attribute{
-			{Key: "testKey", Value: "testValue"},
-		},
+		Kind:       tracing.SpanKindClient,
+		Attributes: []tracing.Attribute{initialAttr},
+		Links:      []tracing.Link{initialLink},
 	})
 	defer func() { endSpan(nil) }()
 
-	require.NotNil(t, tracer.SpanFromContext(ctx))
+	// get the created span from context and add attributes and links
+	// they will get verified with SpanValidator provider
+	spn := tracer.SpanFromContext(ctx)
+	spn.SetAttributes(testAttr)
+	spn.AddLink(testLink)
 }
 
 func TestNewSpanValidator(t *testing.T) {
+	// attributes and links used when starting the span
+	initialAttr := tracing.Attribute{Key: "initialAttrKey", Value: "initialAttrValue"}
+	initialLink := tracing.Link{Attributes: []tracing.Attribute{{Key: "initialLinkKey", Value: "initialLinkValue"}}}
+	// attributes and links added after starting the span
 	testAttr := tracing.Attribute{Key: "testSetAttrKey", Value: "testSetAttrValue"}
-	testLink := tracing.Link{Attributes: []tracing.Attribute{{Key: "testLinkKey", Value: "testLinkValue"}}}
+	testLink := tracing.Link{Attributes: []tracing.Attribute{{Key: "testAddLinkKey", Value: "testAddLinkValue"}}}
 
 	provider := NewSpanValidator(t, SpanMatcher{
 		Name:   "TestSpan",
 		Kind:   tracing.SpanKindClient,
-		Status: tracing.SpanStatusUnset,
-		Attributes: []tracing.Attribute{testAttr,
-			{Key: "testKey", Value: "testValue"},
-		},
-		Links: []tracing.Link{testLink},
+		Status: tracing.SpanStatusError,
+		// error.type is also expected because the span ended with an error
+		Attributes: []tracing.Attribute{initialAttr, testAttr, {Key: "error.type", Value: "*errors.errorString"}},
+		Links:      []tracing.Link{initialLink, testLink},
 	})
 	tracer := provider.NewTracer("module", "version")
 	require.NotNil(t, tracer)
 	require.True(t, tracer.Enabled())
 
 	ctx, endSpan := runtime.StartSpan(context.Background(), "TestSpan", tracer, &runtime.StartSpanOptions{
-		Kind: tracing.SpanKindClient,
-		Attributes: []tracing.Attribute{
-			{Key: "testKey", Value: "testValue"},
-		},
+		Kind:       tracing.SpanKindClient,
+		Attributes: []tracing.Attribute{initialAttr},
+		Links:      []tracing.Link{initialLink},
 	})
 	spn := tracer.SpanFromContext(ctx)
 	spn.SetAttributes(testAttr)
 	spn.AddLink(testLink)
-	defer func() { endSpan(nil) }()
+	defer func() { endSpan(errors.New("test error")) }()
 
 	require.NotNil(t, tracer.SpanFromContext(ctx))
 	require.NotNil(t, tracer.LinkFromContext(ctx))