+++ sdk/internal/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Features Added
 
-# Added function `tracing.New`
+* Added function `tracing.NewSpanValidator` to create a span validator for use with in unit tests.
+* Added type `tracing.SpanMatcher` to match spans in a span validator.
 
 ### Breaking Changes